More changes for A2 (see description)
Inserted the fonts we chose in A1 and set them up. Made the buttons on the homepage link to their respective pages (where available). Made logo link to homepage. Added some extra comments and references. Did a little formatting on About Us page. Made text on hero image bigger and sit better on mobile, and adjusted the sizing of the columns for the drinks on homepage. Changed "Stockist" to "Stockists" in side nav. Changed site title from "ReactApp" to "Feel Good Drinks." Replaced React logos in Instagram section on homepage with random instagram photo for now.