Changes to About Us, Contact Us, Footer, Shop and Privacy Policy
Fiddled with styling on About Us page, and replaced one of the logo images. Fiddled a bit with the footer including positioning the social icons better, and making the copyright smaller. Made the email and phone number clickable on Contact page. Made the Privacy Policy page. Changed the title of Shop from "Shop Page" to "Shop."