Added in changes that Harriet suggested and fiddled with Instagram
Changed the logo to remove the catch line. Changed the image in the tab to the feel good logo. Removed unused pages, changed instagram images, more padding on hero text. Changed button design on homepage, and added fun social icons in footer as well as making footer look better. Made page orange in side nav when you're on that page.