Made some of Harriet's suggested changes
Changed the picture in the Man Behind The Brand section to a man (not sure if it's Ed but closest I could find). Added an image at the top of the About Us page, adjusted size of infographics so that there's more space around as Harriet suggested. Made footer site map links smaller, made footer smaller on bigger screens too.